[Q-hotfix] PaperListItemVo.createUser Integer→Long — 雪花 user_id 容纳
根因：RuoYi 全局 MP idType=ASSIGN_ID（雪花），register 新用户 user_id 是 19 位
雪花数（如 2057993087728254978）。Q 卡 INSERT biz_paper 时 create_by 存这个字符串。
list 查询 SQL CAST(p.create_by AS SIGNED) → BIGINT，但 VO createUser 是 Integer
→ JDBC 取值抛 SQLDataException：'2057993087728254978' is outside of valid range
for type java.lang.Integer，整个 /page 接口炸 500。

修法：createUser Integer → Long。Jackson 全局 BigNumberSerializer 自动将超
SAFE_INTEGER (2^53-1) 的 Long 序列化为 string，FE JS 不丢精度。

回归：Q 4 + U 13 + v1 13 + v05 15 = 45/45 PASS / 0 退化

详 PRD/A-2026-05-22-Q-compose-workbench/PRD.md §10.2

+++ ruoyi-modules/ruoyi-book/src/main/java/org/dromara/book/domain/vo/PaperListItemVo.java
@@ -18,7 +18,7 @@ import java.util.Date;
  *   <li>{@code createTime} STRING 'YYYY-MM-DD' — misikt 真响应是日期串，跟 select 接口 ms timestamp 区别</li>
  *   <li>{@code status} Integer — misikt 真响应 1 整数，不是 '1' 字符串（DB CHAR(1) 转 Integer）</li>
  *   <li>{@code paperType} Integer — misikt 真响应 1 / 2 整数</li>
- *   <li>{@code createUser} Integer — misikt 真响应 2 整数（DB create_by VARCHAR(64) 转 Integer）</li>
+ *   <li>{@code createUser} Long — misikt 真响应小整数（D 卡 ETL admin id）；Q 卡后 RuoYi 雪花 19 位 user_id 需 Long 容纳（详 Q 卡 hotfix 沉淀）</li>
  *   <li>{@code finishTime} 恒 null（DB 无此列 — misikt 真响应也都 null）</li>
  *   <li>{@code hgScore} / {@code directoryName} / {@code frameTextContentId} 大多 null（透传 DB 原值）</li>
  *   <li>{@code sort} Integer — DB INT，通常 = id</li>
@@ -62,8 +62,12 @@ public class PaperListItemVo implements Serializable {
     /** 合格分（DB DECIMAL，misikt 真响应一般 null） */
     private Integer hgScore;
 
-    /** 创建人 id，DB biz_paper.create_by VARCHAR → Integer 对齐 misikt */
-    private Integer createUser;
+    /**
+     * 创建人 id，DB biz_paper.create_by VARCHAR(64) → Long。
+     * D 卡 ETL admin id 是小整数（misikt 真响应口径 Integer），Q 卡 RuoYi 注册 user_id 是雪花 19 位（超 Integer.MAX）。
+     * 改 Long 后老数据小整数兼容、新数据雪花容纳；FE JS Number 接收 >2^53 雪花会丢精度但不影响身份比较 + 显示。
+     */
+    private Long createUser;
 
     /** 目录名（misikt 真响应一般 null） */
     private String directoryName;
